Add files via upload
The training photos are over 100 pictures so I couldn't upload that, so sorry about that. Well, my method to create the training-photos-folders is to use AE (by editing the videos I provided here) and to get .jpeg files into the folders (which are directly named with 'broken' and 'unbroken'). Then you can use this shit-assed-like CNN to get a prediction of the new photo of the beam, however, the results showed0 pretty good, tks again!  There are totally 3 main steps in this project CREATE THE PICKLE FILE (loading_data_color.py), SET MODELS(set_models_flowers_color.py) and also I provide each type of beams for 2 test_pictures and you can use easy_gui_predcition.py to try for the prediction!